backend/requirements.txt
# Django y DRF
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
django-filter==23.3

# Autenticación
dj-rest-auth==5.0.2
django-allauth==0.57.0
djangorestframework-simplejwt==5.3.0

# Base de datos
psycopg2-binary==2.9.7

# Pagos
requests==2.31.0
mercadopago==2.0.0

# Utilidades
Pillow==10.0.1
python-decouple==3.8
celery==5.3.4
redis==5.0.1

# Desarrollo
django-debug-toolbar==4.2.0
django-extensions==3.2.3

# Producción
gunicorn==21.2.0
whitenoise==6.6.0
